+++ src/dao/dslam/impl/metalico/zhone/ZhoneMetalicoComboDslam.java
@@ -7,6 +7,8 @@ package dao.dslam.impl.metalico.zhone;
 
 import br.net.gvt.efika.customer.InventarioRede;
 import dao.dslam.impl.ComandoDslam;
+import dao.dslam.impl.login.LoginLento;
+import dao.dslam.impl.retorno.TratativaRetornoUtil;
 import java.math.BigInteger;
 import java.util.List;
 import model.dslam.consulta.EstadoDaPorta;
@@ -19,8 +21,6 @@ import model.dslam.consulta.metalico.Modulacao;
 import model.dslam.consulta.metalico.TabelaParametrosMetalico;
 import model.dslam.consulta.metalico.TabelaRedeMetalico;
 import model.dslam.credencial.Credencial;
-import dao.dslam.impl.login.LoginLento;
-import dao.dslam.impl.retorno.TratativaRetornoUtil;
 
 /**
  *
@@ -80,12 +80,12 @@ public class ZhoneMetalicoComboDslam extends ZhoneMetalicoDslam {
         List<String> leVlans = this.getCd().consulta(this.getComandoConsultaVlan(i)).getRetorno();
         List<String> leVlanBanda = TratativaRetornoUtil.tratZhone(leVlans, "0-adsl-0-35", "-?\\.?(\\d+((\\.|,| )\\d+)?)");
 
-        BigInteger cvlan = new BigInteger("0");
-        BigInteger p100 = new BigInteger("0");
+        Integer cvlan = new Integer("0");
+        Integer p100 = new Integer("0");
 
         if (leVlanBanda != null) {
-            cvlan = new BigInteger(leVlanBanda.get(1));
-            p100 = new BigInteger(leVlanBanda.get(0));
+            cvlan = new Integer(leVlanBanda.get(1));
+            p100 = new Integer(leVlanBanda.get(0));
         }
         VlanBanda vlanBanda = new VlanBanda(cvlan, p100);
 
@@ -97,12 +97,12 @@ public class ZhoneMetalicoComboDslam extends ZhoneMetalicoDslam {
         List<String> leVlans = this.getCd().consulta(this.getComandoConsultaVlan(i)).getRetorno();
         List<String> leVlanBanda = TratativaRetornoUtil.tratZhone(leVlans, "0-adsl-0-36", "-?\\.?(\\d+((\\.|,| )\\d+)?)");
 
-        BigInteger cvlan = new BigInteger("0");
-        BigInteger p100 = new BigInteger("0");
+        Integer cvlan = new Integer("0");
+        Integer p100 = new Integer("0");
 
         if (leVlanBanda != null) {
-            cvlan = new BigInteger(leVlanBanda.get(1));
-            p100 = new BigInteger(leVlanBanda.get(0));
+            cvlan = new Integer(leVlanBanda.get(1));
+            p100 = new Integer(leVlanBanda.get(0));
         }
         VlanVoip vlanVoip = new VlanVoip(cvlan, p100);
 
@@ -114,12 +114,12 @@ public class ZhoneMetalicoComboDslam extends ZhoneMetalicoDslam {
         List<String> leVlans = this.getCd().consulta(this.getComandoConsultaVlan(i)).getRetorno();
         List<String> leVlanBanda = TratativaRetornoUtil.tratZhone(leVlans, "0-adsl-0-37", "-?\\.?(\\d+((\\.|,| )\\d+)?)");
 
-        BigInteger cvlan = new BigInteger("0");
-        BigInteger p100 = new BigInteger("0");
+        Integer cvlan = new Integer("0");
+        Integer p100 = new Integer("0");
 
         if (leVlanBanda != null) {
-            cvlan = new BigInteger(leVlanBanda.get(1));
-            p100 = new BigInteger(leVlanBanda.get(0));
+            cvlan = new Integer(leVlanBanda.get(1));
+            p100 = new Integer(leVlanBanda.get(0));
         }
         VlanVod vlanVod = new VlanVod(cvlan, p100);
 
@@ -130,10 +130,10 @@ public class ZhoneMetalicoComboDslam extends ZhoneMetalicoDslam {
     public VlanMulticast getVlanMulticast(InventarioRede i) throws Exception {
         List<String> leVlans = this.getCd().consulta(this.getMult(i)).getRetorno();
         List<String> leVlanMult = TratativaRetornoUtil.tratZhone(leVlans, "0-adsl-0-38", "-?\\.?(\\d+((\\.|,| )\\d+)?)");
-        BigInteger cvlan = new BigInteger("0");
+        Integer cvlan = new Integer("0");
 
         if (leVlanMult != null) {
-            cvlan = new BigInteger(leVlanMult.get(0));
+            cvlan = new Integer(leVlanMult.get(0));
         }
         VlanMulticast vlanMult = new VlanMulticast();
         vlanMult.setCvlan(cvlan);
